fix: preserve resumed context for one-shot tasks
