+++ rust/crates/astra-cli/src/tui/mod.rs
@@ -4,9 +4,11 @@ mod testing;
 mod tests;
 
 mod agent_control_status;
+mod agent_view;
 mod app_event;
 pub(crate) mod approval;
 pub(crate) mod background_tasks;
+mod bg_task_proxy;
 mod board_pin;
 mod bottom_pane;
 mod cancel_fanout;
@@ -34,6 +36,7 @@ mod markdown;
 mod markdown_render;
 mod mention_menu;
 pub(crate) mod path_style;
+mod plan_mode;
 mod render;
 mod session_picker;
 mod shimmer;
